added abstract controller for handling multiple controllers
