+++ index.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App;
 
 require_once "src/Utils/debug.php";
-require_once "src/Controller.php";
+require_once "src/NoteController.php";
 require_once "src/Request.php";
 require_once "src/Exception/AppException.php";
 
@@ -22,8 +22,8 @@ $configuration = require_once "config/config.php";
 $request = new Request($_GET,$_POST);
 
 try {
-    Controller::initConfiguration($configuration);
-    $controller = new Controller($request);
+    AbstractController::initConfiguration($configuration);
+    $controller = new NoteController($request);
     $controller->run();
 } catch(ConfigurationException $exception) {
     echo "Something went wrong with configuration";
