+++ src/View.php
@@ -7,7 +7,7 @@ namespace App;
 
 class View
 {
-    public function render(string $page, array $params): void
+    public function render(string $page, array $params = []): void
     {
         require_once("templates/layout.php");
     }
